[jak 2] fix boot, increase level heap sizes (#2252)
Increase level heaps and borrow heaps. The level heap increase was
likely not needed, but better safe than sorry. We allocate the 128 MB
main heap anyway so there's no harm.

Also fix the crash when using `-boot`. As I thought it was just a
one-line typo in the kernel.

+++ common/type_system/TypeSystem.cpp
@@ -648,12 +648,12 @@ MethodInfo TypeSystem::define_method(Type* type,
     int bad_arg_idx = -99;
     // make sure we aren't changing anything that isn't the return type.
     if (!existing_info.type.is_compatible_child_method(ts, type->get_name(), &bad_arg_idx) &&
-        bad_arg_idx != ts.arg_count() - 1) {
+        bad_arg_idx != (int)ts.arg_count() - 1) {
       throw_typesystem_error(
           "The method {} of type {} was originally defined as {}, but has been "
           "redefined as {} (see argument index {})\n",
           method_name, type->get_name(), existing_info.type.print(), ts.print(), bad_arg_idx);
-    } else if (bad_arg_idx == ts.arg_count() - 1 &&
+    } else if (bad_arg_idx == (int)ts.arg_count() - 1 &&
                !tc(existing_info.type.last_arg(), ts.last_arg())) {
       throw_typesystem_error(
           "The method {} of type {} was originally defined as returning {}, but has been redefined "
